# api/openapi.yaml
openapi: 3.1.0
info:
  title: PVM Agent Sandbox API
  version: 1.0.0
  description: |
    Control-plane REST surface of PVM (`uml-container`), a User-Mode Linux
    agent-sandbox manager. Covers the `/api` surface (Bearer auth via
    `API_SECRET`), the E2B-compatible root surface (`X-API-KEY` or Bearer),
    unauthenticated observability endpoints (`/healthz`, `/version`),
    `/metrics`, and the envd Connect-JSON compatibility service on :49983.
    The envd WebSocket version service (:49982) speaks a 5-byte envelope
    frame codec (1 flags + 4 len BE + JSON; end-stream flag 0x02) and is not
    expressible as an OpenAPI path; see the `envd` tag descriptions below.
  license:
    name: Repo license
...
servers:
  - url: http://localhost:8080
    description: API server (agentpvm api)
  - url: http://localhost:3000
    description: WebUI server (agentpvm webui; same routes)
  - url: http://localhost:49983
    description: envd Connect-JSON compatibility (no auth per E2B semantics)
security:
  - bearerAuth: []
...
paths:
...
  /api/tasks/{id}/console:
    get:
      tags: [tasks]
      summary: Recent console output (ring buffer tail)
      parameters: [{ $ref: "#/components/parameters/TaskIdPath" }]
      responses:
        "200":
          description: Console tail
          content:
            application/json:
              schema:
                type: object
                properties:
                  tail: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/tasks/{id}/console/stream:
    get:
      tags: [tasks]
      summary: Server-Sent Events console stream (base64 chunks + ping keepalives)
      description: >-
        Content-Type text/event-stream. The first `event: start` message
        carries {task, offset}; subsequent `data:` lines carry base64-
        encoded console chunks. `: ping` comment lines are keepalives.
      parameters: [{ $ref: "#/components/parameters/TaskIdPath" }]
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/approvals/{id}/edit:
    post:
      tags: [approvals]
      summary: Edit a pending ticket's params (returns it to pending)
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                params:
                  type: object
                  additionalProperties: true
                reason: { type: string }
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }

  # ---------------------------------------------------------------- policy
...
  /api/templates/from-snapshot:
    post:
      tags: [templates]
      summary: Promote a task snapshot into a READY template (backing chain flattened to a standalone raw image)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task]
              properties:
                task: { type: string }
                snapshot_id: { type: string, description: empty = the task's newest snapshot }
                alias: { type: string }
      responses:
        "201": { $ref: "#/components/responses/OkJson" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Error" }
  /api/templates/{id}/inspect:
    get:
      tags: [templates]
      summary: Template record with image size and SHA-256
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "200": { $ref: "#/components/responses/OkJson" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/templates/{id}/preview:
    post:
      tags: [templates]
      summary: Boot an ephemeral sandbox from the template and return its console output
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                command: { type: string }
                timeout_seconds: { type: integer, default: 30, maximum: 180 }
      responses:
        "200": { $ref: "#/components/responses/OkJson" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict", description: template not READY }
        "503": { $ref: "#/components/responses/Error", description: no UML kernel on this host }

  # ---------------------------------------------------------------- egress
...
  /api/network/portmaps:
    get:
      tags: [network]
      summary: List inbound host-port to guest mappings
      responses:
        "200": { $ref: "#/components/responses/OkJson" }
    post:
      tags: [network]
      summary: Publish a host port forwarding to a task's guest (bridge dataplane only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task, host_port, guest_port]
              properties:
                task: { type: string }
                host_port: { type: integer, minimum: 1, maximum: 65535 }
                guest_port: { type: integer, minimum: 1, maximum: 65535 }
                guest_ip: { type: string, description: optional override; defaults to the task's recorded guest IP }
                protocol: { type: string, enum: [tcp, udp], default: tcp }
      responses:
        "201": { $ref: "#/components/responses/OkJson" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict", description: tc-mode tasks cannot map ports }
  /api/network/portmaps/{task}/{hostPort}:
    delete:
      tags: [network]
      summary: Remove one inbound port mapping
      parameters:
        - { $ref: "#/components/parameters/TaskPath" }
        - { name: hostPort, in: path, required: true, schema: { type: integer } }
        - { name: protocol, in: query, schema: { type: string, enum: [tcp, udp], default: tcp } }
      responses:
        "204": { description: Removed }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- identity
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token; the token is the controller's API_SECRET.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: E2B SDK-style API key header; value is the API_SECRET.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Resource id (^[a-zA-Z0-9_-]+$).
    TaskIdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Task id; a unique prefix is accepted (short id).
    TaskPath:
      name: task
      in: path
      required: true
      schema: { type: string }
      description: Task id or unique prefix.
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Error:
      description: Server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    OkJson:
      description: JSON status payload
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable error text.
...
    Ticket:
      type: object
      properties:
        id: { type: string }
        task: { type: string }
        tool: { type: string }
        params:
          type: object
          additionalProperties: true
        status: { type: string, enum: [pending, approved, rejected] }
        deadline: { type: string, format: date-time }
        decided_by: { type: string }
...